% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHNColor.R
\name{CHNColor}
\alias{CHNColor}
\title{A Color Funtion}
\usage{
CHNColor(outPutLength, gradient = FALSE, customColor = c(),
  seed = 25, show = FALSE)
}
\arguments{
\item{outPutLength}{length of the color vector}

\item{gradient}{Do you with to have a color gradient. Defaults to TRUE}

\item{customColor}{Vector of colors you want to be included. Those are appended to the outputVector}

\item{seed}{If you have more than 20 colors the colors are chosen Randomly. To make it reproducible this seed is given. Default is 25.}

\item{show}{switches the output not to a vector of colors but a plot, to visualize the pallet. Defaults to FALSE}
}
\description{
This function allows you to create a vector of distinct colors.
}
\examples{
CHNColor(6)
}
\keyword{colors}
